feat: Add support for AOF file and dump all keys
This commit adds support for an AOF (Append-Only File) in the Redis server. It introduces a new configuration variable `AOFFile` in the `config/main.go` file, which specifies the file path for the AOF. Additionally, a new file `core/aof.go` is added, which contains functions for dumping all keys to the AOF file and rewriting the AOF file. The `DumpAllAOF` function is responsible for iterating over all keys in the store and writing the corresponding SET commands to the AOF file. The `evalBGREWRITEAOF` function is added to evaluate the BGREWRITEAOF command, which triggers the rewriting of the AOF file.