feat: Update eviction strategy to use LRU algorithm
This commit updates the eviction strategy in the Redis server to use the LRU (Least Recently Used) algorithm. It modifies the `core/eviction.go` file to implement the `evictAllkeysLRU` function, which populates an eviction pool based on the last accessed time of keys and evicts the least recently used keys when the number of keys to evict exceeds the configured limit. The `getCurrentClock` function is added to get the current time in Unix format, and the `getIdleTime` function calculates the idle time of keys based on the current time and their last accessed time. This change improves the efficiency of key eviction in the Redis server.