feat: Update .gitignore and add Dockerfile, docker-compose.yml, and docker_run.sh
This commit adds and updates several files related to Docker configuration. The .gitignore file is modified to include the redis_exporter, godocker, and prometheus files. A new Dockerfile is added, which specifies the parent image as golang:1.22.3-bookworm and sets up the app directory, copies the source code, installs Go dependencies, and builds the app. The docker-compose.yml file is also added, which defines services for redis-server, redis-exporter, prometheus-server, and grafana-ui, along with their respective configurations. Lastly, the docker_run.sh script is added to stop and remove any existing containers, build the goredis image, and run a new container with port mapping. These changes enable easier deployment and management of the Redis server using Docker.